refactor : 패키지 위치 변경 전체 게임의 Controller 역할을 할 BaseballGame 클래스의 위치가 domain보다는, baseball 패키지에 속하는 것이 더 옳은 것 같아 위치 변경.
