+++ src/main/java/utils/NumberListGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package utils;
 
 public class NumberListGenerator {
 }
